added add post slice

+++ Server/src/controllers/comment.js
@@ -1,7 +1,7 @@
-const { INTERNAL_SERVER_ERROR, BAD_REQUEST, OK, FORBIDDEN } = require('../constants/statusCodes');
-const { PostModel } = require('../modals/Post');
-const { CommentModel } = require('../modals/Comment');
-const { UserModel } = require('../modals/User');
+const { INTERNAL_SERVER_ERROR, BAD_REQUEST, OK, FORBIDDEN, CREATED } = require('../constants/statusCodes');
+const { PostModel } = require('../models/Post');
+const { CommentModel } = require('../models/Comment');
+const { UserModel } = require('../models/User');
 
 const addComment = async (req, res) => {
     try {
@@ -22,7 +22,7 @@ const addComment = async (req, res) => {
         await UserModel.findByIdAndUpdate(req.user._id, {
             $push: { replies: comment._id }
         });
-        res.status(OK).json({
+        res.status(CREATED).json({
             message: 'comment added successfully!'
         })
     } catch (err) {
@@ -35,10 +35,13 @@ const addComment = async (req, res) => {
 
 const deleteComment = async (req, res) => {
     try {
-        const { commentId } = req.params;
-        if (!commentId) return res.status(BAD_REQUEST).json({ message: 'comment id required!' });
+        const { postId, commentId } = req.params;
+        if (!commentId || !postId) return res.status(BAD_REQUEST).json({ message: 'comment and post id required!' });
         const comment = await CommentModel.findById(commentId);
         if (!comment) return res.status(BAD_REQUEST).json({ message: 'comment not found!' });
+        const post = await PostModel.findById(postId);
+        if (!post) return res.status(BAD_REQUEST).json({ message: 'post not found!' });
+
         if (comment.admin.toString() !== req.user._id.toString()) {
             return res.status(FORBIDDEN).json({
                 message: 'cannot delete another users comment'
@@ -47,9 +50,11 @@ const deleteComment = async (req, res) => {
         await UserModel.findByIdAndUpdate(req.user._id, {
             $pull: { replies: commentId }
         });
-        await PostModel.findByIdAndUpdate(comment.post, {
-            $pull: { comments: commentId }
-        });
+        if(!post.comments.includes(commentId)) return res.status(BAD_REQUEST).json({
+            mesage:'The post does not include the comment'
+        })
+        post.comments.pop(commentId);
+        await post.save();
         await CommentModel.findByIdAndDelete(commentId);
         return res.status(OK).json({
             message: 'comment deleted successfully!'
